Release 20.03

+++ src/Core/BoundedContexts/Headquarters/WB.Core.BoundedContexts.Headquarters/Views/SystemLog/SystemLog.cs
@@ -6,6 +6,6 @@ namespace WB.Core.BoundedContexts.Headquarters.Views.SystemLog
     public class SystemLog
     {
         public IEnumerable<SystemLogItem> Items { get; set; }
-        public long TotalCount { get; set; }
+        public int TotalCount { get; set; }
     }
 }
